requirements.txt
mss
keras
numpy
pillow
keyboard
scikit-learn
